+++ client/src/index.js
@@ -3,19 +3,6 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-import firebase from 'firebase';
-
-// Add here Var config from firebase console-> console.firebase.google.com
-var config={
-  apikey:" ",
-  authDomain:" ",
-  databaseURL:" ",
-  storageBucket:" ",
-};
-
-//Initialize App Config
-firebase.initializeApp(config);
-
 ReactDOM.render(
   <React.StrictMode>
     <App />
